Add badges for `arduino-libraries/Arduino_BHY2` and `arduino-libraries/Arduino_BHY2Host`
These libraries are a little unusual in that they are hosted under the `arduino/nicla-sense-me-fw` repository, then mirrored to dedicated repositories. Because the development work occurs exclusively in the upstream repository, that is where the CI workflow is located. The same workflow is used for both the libraries.

Even though it means each library's badge is a duplicate of the other, I decided it would be less confusing to follow the established format of one row per library rather than merging two library entries into a single row.